test: unify block_number in mocks to u32

+++ pallets/evm/src/mock.rs
@@ -61,7 +61,7 @@ frame_support::construct_runtime! {
 }
 
 parameter_types! {
-    pub const BlockHashCount: u64 = 250;
+    pub const BlockHashCount: u32 = 250;
     pub BlockWeights: frame_system::limits::BlockWeights =
         frame_system::limits::BlockWeights::simple_max(1024);
 }
@@ -71,7 +71,7 @@ impl frame_system::Config for Test {
     type BaseCallFilter = frame_support::traits::Everything;
     type BlockHashCount = BlockHashCount;
     type BlockLength = ();
-    type BlockNumber = u64;
+    type BlockNumber = u32;
     type BlockWeights = ();
     type Call = Call;
     type DbWeight = ();
